Split up unit tests for performance
